notifications: host notifications server
